Move Sidebar and Messenger to the Dashboard component
