+++ whatsapp-client/src/pages/Dashboard/components/Sidebar/Sidebar.css
